Dv  workbranch

See merge request !18

+++ src/core/transactions/transactions/regular/payments/base/BasePaymentTransaction.h
@@ -84,7 +84,6 @@ protected:
         Common_FinalPathConfigurationChecking,
         Common_Recovery,
         Common_ClarificationTransaction
-
     };
 
     enum VotesRecoveryStages {
@@ -95,7 +94,7 @@ protected:
 
 protected:
     // TODO: move it into separate *.h file.
-    typedef uint16_t PathUUID;
+    typedef uint64_t PathUUID;
 
     // Stages handlers
     virtual TransactionResult::SharedConst runVotesCheckingStage();
